feat(benchmark): wire MetricCapability architecture into cold-start benchmarks
sanitizedMetricsForBenchmark()/measureStartupWithCapabilityReport() lived
in benchmark/src/androidTest/.../BenchmarkUtils.kt, which is dead code for
this com.android.test module (only src/main is compiled into the
instrumentation APK) and was never called from anywhere reachable. It also
imported androidx.benchmark.macro.MacrobenchmarkRule instead of the
.junit4 variant actually used by the test classes -- likely why it was
never wired in to begin with.

Move the capability-report logic into the real src/main/BenchmarkUtils.kt
with the import fixed, add the kotlinx-serialization-json dependency the
JSON logging needs (:shared only exposes it as `implementation`, not
transitive), and switch coldStartup_compose()/coldStartup_view() to call
measureStartupWithCapabilityReport() instead of hardcoding a metrics list.
Delete the now-redundant dead file.

Fast-scroll tests and SmokeBenchmark.kt are left as-is; their custom swipe
measureBlock doesn't fit this cold-start-shaped helper.
