+++ benchmark/src/main/java/dev/egarcia/andperf/benchmark/ViewBenchmarks.kt
@@ -2,9 +2,9 @@ package dev.egarcia.andperf.benchmark
 
 import androidx.benchmark.macro.FrameTimingMetric
 import androidx.benchmark.macro.StartupMode
-import androidx.benchmark.macro.StartupTimingMetric
 import androidx.benchmark.macro.junit4.MacrobenchmarkRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import dev.egarcia.andperf.shared.capability.MetricType
 import org.junit.Assume
 import org.junit.Rule
 import org.junit.Test
@@ -22,14 +22,14 @@ class ViewBenchmarks {
         Assume.assumeTrue("Skipping test because target package $pkg is not installed", BenchmarkUtils.isPackageInstalled(pkg))
 
         try {
-            // Match ComposeBenchmarks: use both StartupTimingMetric and FrameTimingMetric
+            // Match ComposeBenchmarks: request both StartupTimingMetric and FrameTimingMetric
             // for a fair apples-to-apples comparison of cold-start and scroll performance.
-            rule.measureRepeated(
+            // The capability probe drops metrics the device can't actually surface.
+            rule.measureStartupWithCapabilityReport(
                 packageName = pkg,
-                metrics = listOf(StartupTimingMetric(), FrameTimingMetric()),
-                iterations = 3,
+                requestedMetrics = listOf(MetricType.STARTUP, MetricType.FRAME_TIMING),
                 startupMode = StartupMode.COLD,
-                measureBlock = { startActivityAndWait() }
+                iterations = 3
             )
         } catch (t: Throwable) {
             // Device may not surface frame metrics — skip gracefully rather than fail.
